+++ src/main/java/com/guinardpouard/imposteur/application/GameService.java
@@ -1,39 +1,68 @@
 package com.guinardpouard.imposteur.application;
 
-import com.guinardpouard.imposteur.domain.model.Room;
-import com.guinardpouard.imposteur.domain.model.WordPair;
+import com.guinardpouard.imposteur.domain.model.*;
+import com.guinardpouard.imposteur.application.port.GamePublisher;
 import com.guinardpouard.imposteur.domain.repository.RoomRepository;
+import com.guinardpouard.imposteur.application.mapper.PrivateRoomUpdateMapper;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
+
+import java.util.Map;
 
 @Service
 public class GameService {
 
+    private final static Logger log = LoggerFactory.getLogger(GameService.class);
     private final RoomRepository roomRepository;
+    private final GamePublisher gamePublisher;
+    private final PrivateRoomUpdateMapper privateRoomUpdateMapper;
 
-    public GameService(RoomRepository roomRepository) {
+    public GameService(RoomRepository roomRepository, GamePublisher gamePublisher,
+                       PrivateRoomUpdateMapper privateRoomUpdateMapper) {
         this.roomRepository = roomRepository;
+        this.gamePublisher = gamePublisher;
+        this.privateRoomUpdateMapper = privateRoomUpdateMapper;
     }
 
-    public Room startGame(String roomId, String hostId) {
+    public void startGame(String roomId, String hostId) {
         Room room = roomRepository.findById(roomId)
                 .orElseThrow(() -> new IllegalArgumentException("Room does not exist"));
 
         WordPair wordPair = new WordPair("Pomme", "Avion");
         room.startGame(hostId, wordPair);
-        roomRepository.save(room);
 
-        return room;
+        for (Map.Entry<PlayerId, PlayerRoundState> state : room.getStates().entrySet()) {
+            gamePublisher.sendWordToPlayer(
+                    privateRoomUpdateMapper.toMessage(
+                            room.getRoomId(),
+                            state.getKey().getValue(),
+                            state.getValue().word()
+                    )
+            );
+        }
+        log.info("Starting game for room {}", room.getRoomId());
+        roomRepository.save(room);
     }
 
-    public Room startNextRound(String roomId, String hostId) {
+    public void startNextRound(String roomId, String hostId) {
         Room room = roomRepository.findById(roomId)
                 .orElseThrow(() -> new IllegalArgumentException("Room does not exist"));
 
         WordPair wordPair = new WordPair("Pomme", "Avion");
-        room.getCurrentGame().startNextRound(wordPair);
-        roomRepository.save(room);
+        room.startNextRound(wordPair);
 
-        return room;
+        for (Map.Entry<PlayerId, PlayerRoundState> state: room.getStates().entrySet()) {
+            gamePublisher.sendWordToPlayer(
+                    privateRoomUpdateMapper.toMessage(
+                            room.getRoomId(),
+                            state.getKey().getValue(),
+                            state.getValue().word()
+                    )
+            );
+        }
+        log.info("Starting next round for room {}", room.getRoomId());
+        roomRepository.save(room);
     }
 
 }
