+++ src/main/java/com/guinardpouard/imposteur/application/event/PlayerView.java
@@ -1,0 +1,7 @@
+package com.guinardpouard.imposteur.application.event;
+
+public record PlayerView(
+        String playerId,
+        String playerName
+) {
+}
