+++ src/main/java/com/guinardpouard/imposteur/application/event/PrivateRoomUpdatedMessage.java
@@ -1,0 +1,8 @@
+package com.guinardpouard.imposteur.application.event;
+
+public record PrivateRoomUpdatedMessage(
+        String roomId,
+        String playerId,
+        String word
+) {
+}
