+++ src/main/java/com/guinardpouard/imposteur/application/event/RoomUpdatedMessage.java
@@ -1,0 +1,10 @@
+package com.guinardpouard.imposteur.application.event;
+
+import java.util.List;
+
+public record RoomUpdatedMessage(
+        String roomId,
+        String roomName,
+        List<PlayerView> playerViewList
+) {
+}
