+++ src/main/java/com/guinardpouard/imposteur/domain/model/FinishedState.java
@@ -15,14 +15,7 @@ public class FinishedState implements GameState {
 
     @Override
     public GameState start(Room room, List<Player> players, String hostId, WordPair wordPair) {
-        if (room.getPlayers().size() < 2) {
-            throw new IllegalStateException("Game cannot start with less than 2 players");
-        }
-        if (!room.isHost(hostId)) {
-            throw new IllegalStateException("Game can be started only by host");
-        }
-
-        room.setCurrentGame(new GameSession(players, wordPair));
+        room.start(players, hostId, wordPair);
         return InProgressState.INSTANCE;
     }
