+++ src/main/java/com/guinardpouard/imposteur/infrastructure/websocket/entrypoint/dto/CreateRoomMessage.java
@@ -1,0 +1,6 @@
+package com.guinardpouard.imposteur.infrastructure.websocket.entrypoint.dto;
+
+public record CreateRoomMessage(
+        String roomName
+) {
+}
