+++ src/main/java/com/guinardpouard/imposteur/infrastructure/websocket/entrypoint/dto/JoinRoomMessage.java
@@ -1,0 +1,7 @@
+package com.guinardpouard.imposteur.infrastructure.websocket.entrypoint.dto;
+
+public record JoinRoomMessage(
+        String roomId,
+        String playerName
+) {
+}
