+++ src/main/java/com/guinardpouard/imposteur/infrastructure/websocket/entrypoint/dto/StartGameMessage.java
@@ -1,0 +1,7 @@
+package com.guinardpouard.imposteur.infrastructure.websocket.entrypoint.dto;
+
+public record StartGameMessage(
+        String roomId,
+        String hostId
+) {
+}
