--- src/main/java/com/guinardpouard/imposteur/websocket/dto/PlayerView.java
@@ -1,7 +1,0 @@
-package com.guinardpouard.imposteur.websocket.dto;
-
-public record PlayerView(
-        String playerId,
-        String playerName
-) {
-}
