--- src/main/java/com/guinardpouard/imposteur/websocket/dto/PrivateRoomUpdatedMessage.java
@@ -1,8 +1,0 @@
-package com.guinardpouard.imposteur.websocket.dto;
-
-public record PrivateRoomUpdatedMessage(
-        String roomId,
-        String roomName,
-        String word
-) {
-}
