--- src/main/java/com/guinardpouard/imposteur/websocket/dto/RoomCreateMessage.java
@@ -1,7 +1,0 @@
-package com.guinardpouard.imposteur.websocket.dto;
-
-public record RoomCreateMessage(
-        String roomName,
-        String playerName
-) {
-}
