--- src/main/java/com/guinardpouard/imposteur/websocket/dto/RoomCreatedMessage.java
@@ -1,7 +1,0 @@
-package com.guinardpouard.imposteur.websocket.dto;
-
-public record RoomCreatedMessage(
-        String roomId,
-        String roomName
-) {}
-
