--- src/main/java/com/guinardpouard/imposteur/websocket/dto/RoomJoinMessage.java
@@ -1,7 +1,0 @@
-package com.guinardpouard.imposteur.websocket.dto;
-
-public record RoomJoinMessage(
-        String roomId,
-        String playerName
-) {
-}
