--- src/main/java/com/guinardpouard/imposteur/websocket/dto/RoomUpdatedMessage.java
@@ -1,10 +1,0 @@
-package com.guinardpouard.imposteur.websocket.dto;
-
-import java.util.List;
-
-public record RoomUpdatedMessage(
-        String roomId,
-        String roomName,
-        List<PlayerView> playerViewList
-) {
-}
