--- src/main/java/com/guinardpouard/imposteur/websocket/dto/StartGameMessage.java
@@ -1,7 +1,0 @@
-package com.guinardpouard.imposteur.websocket.dto;
-
-public record StartGameMessage(
-        String roomId,
-        String hostId
-) {
-}
