+++ src/test/java/com/guinardpouard/imposteur/application/RoomServiceTests.java
@@ -1,7 +1,10 @@
 package com.guinardpouard.imposteur.application;
 
 import com.guinardpouard.imposteur.domain.model.Room;
+import com.guinardpouard.imposteur.application.port.GamePublisher;
 import com.guinardpouard.imposteur.domain.repository.RoomRepository;
+import com.guinardpouard.imposteur.application.mapper.RoomUpdatedMapper;
+import com.guinardpouard.imposteur.application.event.RoomUpdatedMessage;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -13,7 +16,7 @@ import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
 public class RoomServiceTests {
@@ -21,32 +24,41 @@ public class RoomServiceTests {
     private RoomService roomService;
     @Mock
     private RoomRepository mockRoomRepository;
+    @Mock
+    private GamePublisher mockGamePublisher;
+    @Mock
+    private RoomUpdatedMapper mockMapper;
 
     @BeforeEach
     public void setup() {
-        roomService = new RoomService(mockRoomRepository);
+        roomService = new RoomService(mockRoomRepository, mockGamePublisher, mockMapper);
     }
 
     @Test
     void createRoom_shouldCreate_a_new_room() {
-        Room room = roomService.createRoom("room 123", "user-123");
-        assertThat(room).isNotNull();
-        assertThat(room.getPlayers()).isEmpty();
-        assertThat(room.getRoomId()).isNotNull();
-        assertThat(room.getRoomName()).isEqualTo("room 123");
+        when(mockMapper.toMessage(any(Room.class))).thenCallRealMethod();
+        roomService.createRoom("room 123", "user-123");
+
+        verify(mockRoomRepository, times(1))
+                .save(any(Room.class));
+        verify(mockGamePublisher, times(1))
+                .publishRoomCreated(any(RoomUpdatedMessage.class));
     }
 
     @Test
     void addPlayerToRoom_should_add_player_and_return_updated_room() {
         Room room = new Room("room 1", "user-123");
         when(mockRoomRepository.findById(room.getRoomId())).thenReturn(Optional.of(room));
+        when(mockMapper.toMessage(any(Room.class))).thenCallRealMethod();
 
-        room = roomService.addPlayerToRoom("user-456", room.getRoomId(), "player2");
-        assertThat(room).isNotNull();
-        assertThat(room.getPlayers()).hasSize(1);
-        assertThat(room.getPlayers().getFirst().getName()).isEqualTo("player2");
-        assertThat(room.getPlayers().getFirst().getConnectionId()).isEqualTo("user-456");
-        assertThat(room.getPlayers().getFirst().id()).isNotNull();
+        roomService.addPlayerToRoom("user-456", room.getRoomId(), "player2");
+
+        verify(mockRoomRepository, times(1))
+                .findById(eq(room.getRoomId()));
+        verify(mockRoomRepository, times(1))
+                .save(any(Room.class));
+        verify(mockGamePublisher, times(1))
+                .publishRoomJoined(any(RoomUpdatedMessage.class));
     }
 
     @Test
