+++ src/test/java/com/guinardpouard/imposteur/infrastructure/websocket/entrypoint/dto/StartGameMessageTests.java
@@ -1,0 +1,19 @@
+package com.guinardpouard.imposteur.infrastructure.websocket.entrypoint.dto;
+
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class StartGameMessageTests {
+
+    @Test
+    void should_contains_dto() {
+        StartGameMessage msg = new StartGameMessage(
+                "room id",
+                "hostConnectionId"
+        );
+
+        assertThat(msg.roomId()).isEqualTo("room id");
+        assertThat(msg.hostId()).isEqualTo("hostConnectionId");
+    }
+}
