--- src/test/java/com/guinardpouard/imposteur/websocket/dto/PlayerViewTests.java
@@ -1,15 +1,0 @@
-package com.guinardpouard.imposteur.websocket.dto;
-
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-class PlayerViewTests {
-
-    @Test
-    void playerView_should_contains_id_and_name() {
-        PlayerView playerView = new PlayerView("playerId", "playerName");
-        assertThat(playerView.playerId()).isEqualTo("playerId");
-        assertThat(playerView.playerName()).isEqualTo("playerName");
-    }
-}
